# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Islamic LMS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - in: query
          name: role
          schema:
            type: string
        - in: query
          name: gender
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

    post:
      operationId: createUser
      tags: [users]
      summary: Create a user (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/verify:
    patch:
      operationId: verifyStudent
      tags: [users]
      summary: Verify a student and assign federation code
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyStudentBody"
      responses:
        "200":
          description: Student verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # COURSES
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, teacher, female_teacher, qari, male_student, female_student]
        gender:
          type: string
          enum: [male, female]
        section:
          type: ["string", "null"]
          enum: [boys, girls, null]
        level:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, verified, suspended, trial, active]
        federationCode:
          type: ["string", "null"]
        cnicNumber:
          type: ["string", "null"]
        trialStartDate:
          type: ["string", "null"]
        trialEndDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - fullName
        - email
        - role
        - gender
        - status
        - createdAt
        - updatedAt
    CreateUserBody:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, teacher, female_teacher, qari, male_student, female_student]
        gender:
          type: string
          enum: [male, female]
        section:
          type: ["string", "null"]
          enum: [boys, girls, null]
        level:
          type: ["integer", "null"]
        cnicNumber:
          type: ["string", "null"]
      required:
        - fullName
        - email
        - password
        - role
        - gender
    UpdateUserBody:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        level:
          type: ["integer", "null"]
        section:
          type: ["string", "null"]
          enum: [boys, girls, null]
        status:
          type: string
          enum: [pending, verified, suspended, trial, active]
    VerifyStudentBody:
      type: object
      properties:
        federationCode:
          type: string
        status:
          type: string
          enum: [verified, active]
      required:
        - federationCode
        - status
...